Move example nodes to example_nodes folder and import from there. Add 2 nodes TickTimeNode and TextFileInputNode. Update example_math_nodes.py.
