+++ example_math_nodes.py
@@ -1,157 +1,10 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
-import math
-import inspect
-from functools import partial
-
-# add basic math functions to math library
-math.add = lambda x, y: x + y
-math.sub = lambda x, y: x - y
-math.mul = lambda x, y: x * y
-math.div = lambda x, y: x / y
-
-# Transform float to functions
-for constant in ['pi', 'e', 'tau', 'inf', 'nan']:
-    setattr(math, constant, partial(lambda x: x, getattr(math, constant)))
-
-
+import example_nodes as Nodes
 from NodeGraphQt import (NodeGraph,
                          BaseNode,
                          setup_context_menu)
 from NodeGraphQt import QtWidgets, QtCore, PropertiesBinWidget, NodeTreeWidget
-
-
-class DataInputNode(BaseNode):
-    """
-    Input node data.
-    """
-
-    __identifier__ = 'com.chantasticvfx'
-    NODE_NAME = 'Input Numbers'
-
-    def __init__(self):
-        super(DataInputNode, self).__init__()
-        self.add_output('out')
-        self.add_text_input('out', 'Data Output', text='0.4', tab='widgets')
-        self.view.widgets['out'].value_changed.connect(self.update_streams)
-
-
-class MathFunctionsNode(BaseNode):
-    """
-    Math functions node.
-    """
-
-    # set a unique node identifier.
-    __identifier__ = 'com.chantasticvfx'
-
-    # set the initial default node name.
-    NODE_NAME = 'Functions node'
-
-    mathFuncs = [func for func in dir(math) if not func.startswith('_')]
-
-    def __init__(self):
-        super(MathFunctionsNode, self).__init__()
-        self.set_color(25, 58, 51)
-        self.add_combo_menu('funcs', 'Functions', items=self.mathFuncs,
-                            tab='widgets')
-
-        # switch math function type
-        self.view.widgets['funcs'].value_changed.connect(self.addFunction)
-        self.view.widgets['funcs'].value_changed.connect(self.update_streams)
-        self.add_output('output')
-        self.create_property('output', None)
-        self.trigger_type = 'no_inPorts'
-
-        self.view.widgets['funcs'].widget.setCurrentIndex(2)
-
-    def addFunction(self, prop, func):
-        """
-        Create inputs based on math functions arguments.
-        """
-        self.func = getattr(math, func)
-        dataFunc = inspect.getfullargspec(self.func)
-
-        for arg in dataFunc.args:
-            if not self.has_property(arg):
-                inPort = self.add_input(arg)
-                inPort.trigger = True
-                self.create_property(arg, None)
-
-        for inPort in self._inputs:
-            if inPort.name() in dataFunc.args:
-                if not inPort.visible():
-                    inPort.set_visible(True)
-            else:
-                inPort.set_visible(False)
-
-    def run(self):
-        """
-        Evaluate all entries, pass them as arguments of the
-        chosen mathematical function.
-        """
-        for to_port in self.input_ports():
-            if to_port.visible() == False:
-                continue
-            from_ports = to_port.connected_ports()
-            if not from_ports:
-                raise Exception('Port %s not connected!' % to_port.name(),
-                                to_port)
-
-            for from_port in from_ports:
-                from_port.node().run()
-                data = from_port.node().get_property(from_port.name())
-                self.set_property(to_port.name(), float(data))
-
-        try:
-            # Execute math function with arguments.
-            data = self.func(*[self.get_property(inport.name()) for inport in self._inputs if inport.visible()])
-
-            self.set_property('output', data)
-        except KeyError as error:
-            print("An input is missing! %s" % str(error))
-        except TypeError as error:
-            print("Error evaluating function: %s" % str(error))
-
-    def on_input_connected(self, to_port, from_port):
-        """Override node callback method."""
-        self.set_property(to_port.name(), from_port.node().run())
-        self.update_streams()
-
-    def on_input_disconnected(self, to_port, from_port):
-        """Override node callback method."""
-        self.set_property('output', None)
-        self.update_streams()
-
-
-class DataViewerNode(BaseNode):
-    __identifier__ = 'com.chantasticvfx'
-    NODE_NAME = 'Result View'
-
-    def __init__(self):
-        super(DataViewerNode, self).__init__()
-        self.inPort = self.add_input('data')
-        self.add_text_input('data', 'Data Viewer', tab='widgets')
-
-    def run(self):
-        """Evaluate input to show it."""
-        for source in self.inPort.connected_ports():
-            from_node = source.node()
-            try:
-                from_node.run()
-            except Exception as error:
-                print("%s no inputs connected: %s" % (from_node.name(), str(error)))
-                self.set_property('data', None)
-                return
-            value = from_node.get_property(source.name())
-            self.set_property('data', str(value))
-
-    def on_input_connected(self, to_port, from_port):
-        """Override node callback method"""
-        self.run()
-
-    def on_input_disconnected(self, to_port, from_port):
-        """Override node callback method"""
-        self.set_property('data', None)
 
 
 if __name__ == '__main__':
@@ -187,38 +40,41 @@ if __name__ == '__main__':
     graph.node_double_clicked.connect(show_nodes_list)
 
     # registered nodes.
-    reg_nodes = [DataInputNode, DataViewerNode, MathFunctionsNode]
+    reg_nodes = [
+        Nodes.__dict__[n] for n in Nodes.__dict__
+        if hasattr(Nodes.__dict__[n], 'NODE_NAME')
+    ]
 
     for n in reg_nodes:
         graph.register_node(n)
 
-    mathNodeA = graph.create_node('com.chantasticvfx.MathFunctionsNode',
+    mathNodeA = graph.create_node('Math.MathFunctionsNode',
                                 name='Math Functions A',
                                 color='#0a1e20',
                                 text_color='#feab20',
                                 pos=[-250, 70])
 
-    mathNodeB = graph.create_node('com.chantasticvfx.MathFunctionsNode',
+    mathNodeB = graph.create_node('Math.MathFunctionsNode',
                                 name='Math Functions B',
                                 color='#0a1e20',
                                 text_color='#feab20',
                                 pos=[-250, -70])
 
-    mathNodeC = graph.create_node('com.chantasticvfx.MathFunctionsNode',
-                                name='Math Functions C',
-                                color='#0a1e20',
-                                text_color='#feab20',
-                                pos=[0, 0])
+    mathNodeC = graph.create_node('Math.MathFunctionsNode',
+                                  name='Math Functions C',
+                                  color='#0a1e20',
+                                  text_color='#feab20',
+                                  pos=[0, 0])
 
-    inputANode = graph.create_node('com.chantasticvfx.DataInputNode',
+    inputANode = graph.create_node('Inputs.DataInputNode',
                                    name='Input A',
                                    pos=[-500, -50])
 
-    inputBNode = graph.create_node('com.chantasticvfx.DataInputNode',
+    inputBNode = graph.create_node('Inputs.DataInputNode',
                                    name='Input B',
                                    pos=[-500, 50])
 
-    outputNode = graph.create_node('com.chantasticvfx.DataViewerNode',
+    outputNode = graph.create_node('Viewers.DataViewerNode',
                                    name='Output',
                                    pos=[250, 0])
